+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v0.2.0-dev
 ### Added
 * Optionally expand merchant info when fetching transactions
+* Add missing fields on Mondo.Transaction: attachments, local_amount, local_currency
 
 ### Changed
 * Bump poison to 2.0.1